✅ Add unit tests
